+++ psyplot/data.py
@@ -646,6 +646,8 @@ class CFDecoder(object):
             if 'coordinates' in obj.attrs:
                 extra_coords.update(obj.attrs['coordinates'].split())
                 obj.encoding['coordinates'] = obj.attrs.pop('coordinates')
+            if 'grid_mapping' in obj.attrs:
+                extra_coords.add(obj.attrs['grid_mapping'])
             if 'bounds' in obj.attrs:
                 extra_coords.add(obj.attrs['bounds'])
         if gridfile is not None and not isinstance(gridfile, xr.Dataset):
@@ -915,16 +917,26 @@ class CFDecoder(object):
         if not coord_names:
             return
         ret = []
+        matched = []
         for coord in map(lambda dim: coords[dim], filter(
                 lambda dim: dim in coords, chain(
                     coord_names, var.dims))):
             # check for the axis attribute or whether the coordinate is in the
             # list of possible coordinate names
-            if (coord.name not in (c.name for c in ret) and
-                    (coord.attrs.get('axis', '').lower() == axis or
-                     coord.name in getattr(self, axis))):
-                ret.append(coord)
-        if ret:
+            if coord.name not in (c.name for c in ret):
+                if coord.name in getattr(self, axis):
+                    matched.append(coord)
+                elif coord.attrs.get('axis', '').lower() == axis:
+                    ret.append(coord)
+        if matched:
+            if len(matched) > 1:
+                warn("Found multiple matches for %s coordinate in the "
+                     "coordinates: %s. I use %s" % (
+                         axis, ', '.join([c.name for c in matched]),
+                         matched[0].name),
+                     PsyPlotRuntimeWarning)
+            return matched[0]
+        elif ret:
             return None if len(ret) > 1 else ret[0]
         # If the coordinates attribute is specified but the coordinate
         # variables themselves have no 'axis' attribute, we interpret the